move __debug_sender into its own header
